+++ apps/widget-configurator/src/components/ui/buttons/button/Button.styles.ts
@@ -1,0 +1,15 @@
+import {
+  labelledButtonTypographySx,
+  mediumButtonSizeSx,
+  mediumEndIconSpacingSx,
+  sharedBorderSx,
+} from '../base/BaseButton.styles'
+
+import type { SxProps, Theme } from '@mui/material/styles'
+
+export const buttonSx = {
+  ...sharedBorderSx,
+  ...labelledButtonTypographySx,
+  ...mediumButtonSizeSx,
+  ...mediumEndIconSpacingSx,
+} satisfies SxProps<Theme>