+++ libs/widget-lib/src/index.ts
@@ -1,5 +1,11 @@
 export { createCowSwapWidget } from './cowSwapWidget'
-export { WIDGET_IFRAME_ALLOW, WIDGET_IFRAME_REFERRER_POLICY, WIDGET_IFRAME_SANDBOX } from './cowSwapWidget.constants'
+export {
+  WIDGET_CONTAINER_ID,
+  WIDGET_IFRAME_ALLOW,
+  WIDGET_IFRAME_ID,
+  WIDGET_IFRAME_REFERRER_POLICY,
+  WIDGET_IFRAME_SANDBOX,
+} from './cowSwapWidget.constants'
 export type { CowSwapWidgetHandler } from './cowSwapWidget'
 export { resolveFlexibleConfigValues } from './resolveFlexibleConfigValues'
 export { sanitizeWidgetBaseUrl, SHOULD_THROW_IF_INVALID_URL } from './urlUtils'