+++ libs/ui/CHANGELOG.md
@@ -1,5 +1,57 @@
 # Changelog
 
+## [3.9.0](https://github.com/cowprotocol/cowswap/compare/ui-v3.8.1...ui-v3.9.0) (2026-06-25)
+
+### Minor Changes
+
+- feat: Widget Configurator Revamp (#7604)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/assets@2.5.0
+  - @cowprotocol/common-hooks@3.5.0
+  - @cowprotocol/common-utils@3.7.0
+  - @cowprotocol/analytics@3.5.2
+
+## [3.8.1](https://github.com/cowprotocol/cowswap/compare/ui-v3.8.0...ui-v3.8.1) (2026-06-23)
+
+### Patch Changes
+
+- fix: fix near bridging (#7716)
+
+- Updated dependencies []:
+  - @cowprotocol/analytics@3.5.1
+  - @cowprotocol/common-const@3.5.1
+  - @cowprotocol/common-hooks@3.4.1
+  - @cowprotocol/common-utils@3.6.1
+  - @cowprotocol/core@3.5.1
+  - @cowprotocol/currency@1.2.1
+  - @cowprotocol/types@4.5.1
+
+## [3.8.0](https://github.com/cowprotocol/cowswap/compare/ui-v3.7.1...ui-v3.8.0) (2026-06-23)
+
+### Minor Changes
+
+- feat: adjust dApp configuration for Solana as supported chain id (#7525)
+
+- feat(bridge): enable sell=buy feature (#7633)
+
+### Patch Changes
+
+- fix: BTC bridging (#7532)
+
+- fix(explorer): solver info displaying (#7667)
+
+- Updated dependencies []:
+  - @cowprotocol/analytics@3.5.0
+  - @cowprotocol/common-const@3.5.0
+  - @cowprotocol/common-hooks@3.4.0
+  - @cowprotocol/common-utils@3.6.0
+  - @cowprotocol/core@3.5.0
+  - @cowprotocol/currency@1.2.0
+  - @cowprotocol/types@4.5.0
+
 ## [3.7.1](https://github.com/cowprotocol/cowswap/compare/ui-v3.7.0...ui-v3.7.1) (2026-06-09)
 
 ### Patch Changes